Set old object on initial focus load operation
When starting a clockwork, initial focus load operation is executed.
Previously, we have updated only current & new state of the focus,
leaving old state as it came from the external caller. However, the
focus is often incomplete (e.g. without jpegPhoto) at that time.

Now we overwrite also the old state - but only on the first loading.

Related change:

- A safety check was added: If a mapping is to be executed on an
incomplete item, an exception is thrown. It is better to see the issue
immediately than to incur data quality problems detected potentially
much later.

Unrelated change:

- When serializing operation result (for task) we employ less strict
mode, allowing unserializable values to be stored as strings, and also
escaping invalid XML characters. This can be useful e.g. when tracing
is enabled, and is consistent with the new (native) repo behavior.