midpoint-war separated from admin-gui, which is now plain JAR module
- In IDEA the project can still be run from admin-gui module using the
same main class as before.
- Some of the previously provided modules that should go to the JAR
(like Tomcat stuff) is not provided in admin-gui JAR anymore, but is
marked as provided in midpoint-war only (which should soon go away).
- JDBC drivers for Oracle and MS SQL are still provided in admin-gui
for developers, but not distributed, as before.
The POM "code" to exclude them, however, is now not needed in the
separate module.
- Banners are added to midpont-jar+war modules and Maven-filtered to
provide accurate git+other info; admin-gui doesn't need it but we have
it there so the devs don't get scared by different banner.