Move advanced expression support to repo-common
We decided to move resource object classification and (maybe) tag
generation from model-impl to provisioning-impl. In order to do this,
some parts of expression support need to be present in lower layers
(repo-common). Also, SystemObjectCache needs to be there.

So:

1. ModelExpressionThreadLocalHolder was changed to
ExpressionEnvironmentThreadLocalHolder and moved downwards.

2. ExpressionEnvironment was split into original class (reduced to
contain only task + result) and ModelExpressionEnvironment. The former
is now in repo-common, the latter in model-common.

3. Utility methods like evaluateExpressionInContext were moved
from ModelExpressionThreadLocalHolder to ExpressionUtil.

4. SystemObjectCache was pushed down to repo-common.

5. AbstractIntegrationTest#invalidateSystemObjectsCache was removed.
Its functionality is covered by standard cache invalidation mechanism
that is applied when system configuration object is modified.

Unrelated:

- Fixed NPE in ResourceOperationalButtonsPanel.