Move shadow classification to provisioning-impl
This (long-awaited) move was necessary because of the planned
implementation of "smart classification" based on object type
delineation.

The shadow tag generation was moved as well. On the other hand,
synchronization sorter evaluation remains in the model-impl (for now).

Also fixed a couple of POM dependencies.

Work in progress. Some tests may fail.