Unify "no schema" error messages
(At least to some extent.) The idea is that complete schema is always
present if there's a raw schema. So if complete or raw schema is
missing, this is most probably a configuration issue: like no schema
capability declared, or missing static schema if a resource does not
support schema at all, etc. The exception is when the resource is not
complete because it's e.g. unreachable.