Add a concept of refined object class definitions
In order to allow selected refinements to be applied to the whole object
class on a given resource, here we introduce so-called refined
object class definition. It is configured in new `objectClass`
item in `schemaHandling` (alongside existing `objectType`). It is
just like an abstract object type, but it has no kind nor intent,
that's why it is _not_ an object type: it's a refined object class.

The primary motivation was to cleanly resolve the issue of handling
object-class-scoped operations (typically a search), i.e. those with
the query in the form of "objectClass=X" (no kind/intent). We often
need the refined definition to execute such queries (e.g. to know
attributesToGet), but it cannot be the _type_ definition. Or, well,
it may be, but at the cost of ugly hacking with "default" flags, as
it was done previously.

Work in progress. Implementation is not complete. Tests may fail.

More specifically,

1. Added schemaHandling/objectClass item in XSD.
2. Added "refined" state of ResourceObjectClassDefinitionImpl, with
corresponding changes (i.e. moving methods upwards) in the
AbstractResourceObjectDefinitionImpl sub-hierarchy. Adapted the code
in provisioning-impl module.
3. Added ProvisioningContext#wholeClass flag, to distinguish between
type and class-scoped bulk (search) operations.
4. Attempted to fix abstract delegator classes for subclasses
of ResourceObjectDefinition.

Unrelated changes:

- Some cleanups in the test code, e.g. related to constants for
class/attribute names, qualified names, and paths. We utilize the fact
that all resources have the same namespace of "ri" now.
- Fixed argument for ProvisioningService#synchronize method from
ResourceShadowDiscriminator to ResourceShadowCoordinates.
- Fixed a bug in shared ResourceObjectPattern filter being rewritten
when its expressions are evaluated.