Re-allow null intent for dependency specification
Making indent in dependency configuration obligatory
(in d62cd32e87ca66c033e182a73ce8b705eeb872e9) was not a good idea.

This commit implements handling of null intent there: it is interpreted
as "default in a kind". This means that specialized methods for
dependency lookup have been introduced in DependencyProcessor class.

Also, ResourceObjectTypeDependencyType has been separated from
ShadowDiscriminatorType. Not all fields are relevant, and the
meaning (hence the docs) is different as well.