Apply all sorter-provided classifications
During recent changes in synchronization code (most probably in
5e841416523a6d46431443d36b056e91ada0b9f1) the handling of sync sorter
result was broken: it was ignored if the shadow was already classified.

This commit brings back the original behavior: if the sorter provides
a classification, it is applied regardless of whether the shadow was
already classified. (Assuming that the synchronization is not disabled.)

Other changes:

- If the sorter returns none or incomplete classification, the default
classification algorithm is carried out.

- A couple of minor code improvements have been done, e.g. ShadowUpdater
was created to help SynchronizationServiceImpl with saving sync
metadata.

Work in progress.