Remove support for filter clauses in delineation
Instead of having both "overwritable" single-valued `filter` and
"non-overwritable" (cumulative) multivalued `filterClause` delineation
properties, we have simplified the schema to have only cumulative
multivalued `filter`.

This is consistent with the intuition that a subtype should present
a subset of all objects covered by the supertype. So, when going from
a supertype to a subtype we should be allowed only to add restrictions,
not remove/change them.