Improve archetype processing in <objectType> def
When focus/archetypeRef is specified for a resource object type, it is
now enforced to all focuses that have this resource object type linked
(i.e. not only to newly-created ones). See the XSD docs.

Related changes:

- The algorithm for determining archetypes for objects being changed
(by clockwork) was improved a bit.

- Focus template is now present in LensFocusContext (where it belongs),
not in LensContext. Only the externally-set template OID is kept in
LensContext.

- Methods for updating archetypes, archetype policy, and object template
for focus contexts were streamlined.
(See ContextLoader#updateArchetypePolicyAndRelatives.)

- When system configuration is missing, the archetype policies
determination was originally skipped. This is now fixed.

Work in progress.