Improve classification algorithm
The classification now works in two phases: (1) "ordered" type
definitions are tried, and the first matching one is used;
(2) all the others are tried, and either the one marked as "default"
is used, or a heuristic to select the single one is executed.

Also, baseContextClassificationUse configuration option has been
introduced.

Other changes:

1. FindDefinitionForObjectClass replaced with findObjectClassDefinition
(mainly for auxiliary object classes).

2. SynchronizationPolicy now requires object type definition (not only
generic ResourceObjectDefinition)! This means that it's no longer
possible to have <synchronization> section without corresponding
object type definition.

This is an !!! INCOMPATIBLE CHANGE => to RELEASE NOTES !!!