Fix duplicate name delta during shadow refresh
Delta for $shadow/name was created twice: first in
RefreshHelper#refreshShadowAsyncStatus, and then in
ShadowUpdater#extractRepoShadowChanges. This caused the consistency
checks in repository to fail.

The latter method was adapted to allow only single `name` delta
to pass through.

This should resolve MID-7985 and MID-7924.

(cherry-picked and adapted from commit
0de9509050016c7183d18b686f986ac6d865ed89)