Add pieces of embedded objects support
The goal is to support embedding prism object values right in prism
containers. Currently, object values can be embedded only in prism
properties.

This commit contains some pieces of the required functionality, but
it is far from complete:

1. Schema parsing: new "embeddedObject" annotation has been devised
to mark an embedded object (to distinguish it from, e.g., a legacy form
of specifying full object along with the reference - link/linkRef).

2. Data parsing: abstract containers couldn't be parsed, even if
xsi:type declaration was present. (This is not specific to embedded
objects.) A partial hack is present here. But the parsing/serialization
code would deserve a complete rework.

3. Data storage: AbstractMutableObjectable assumes that the container
is always PrismObject/PrismObjectValue combo. We need to support PC/PCV
as well.

Work in progress. Tests are not meant to succeed now.