Handle temporary PCV IDs correctly
Temporary PCV IDs are needed in cases when we create a new container
value and then evaluate some mappings (or swallow secondary deltas
created by other means) that need to reference content of this value.

To do so, we create temporary PCV IDs (e.g. -1, -2, and so on) that
live only during clockwork processing, and are removed just before
the summary delta is being applied.

Currently, this is needed for mappings targeting new
identities/identity[xxx] values.