SelectorOptions: added new hasToFetchPathNotRetrievedByDefault
Previous hasToLoadPath was renamed to hasToIncludePath, because it does
not completely answer the question about loading (but it does about
including).
Version with auto-magic default was removed, calling code always knows.
New hasToFetchPathNotRetrievedByDefault was introduces with very similar
semantics, but more targeted for repo usages.
This new method is now used for all the relevant cases.