AssignmentPanel - Use modified targetRef for concrete panels
Some panels are concretized by targetRef/targetType. In this panels
we can safely assume that targetRef is of static type, which allows
us to create temporary schema overlay with narrowed definition of targetRef.

This narrowed definition is used in collumns models and in search panel
to allow using concrete extension items of Role, Org or Service to serve as
columns and/or basic search fields.

This change also provides nice feature for Axiom Query language, that
user in this specific panels does not need to specify type when dereferencing.