Fix handling exceptions in search post-processing
When an exception occurred during "applyDefinition" for provisioning
objects (resources, shadows) after they were fetched from the
repository, it was not handled correctly. It caused the whole
search operation to fail.

This is now fixed.

However, it is questionable if we want to support this mode of operation
(i.e. searching for more object types e.g. resources + connectors
in a single query) in the long term.