Fix authorization evaluation for the EXISTS filter
When determining the list of items required to evaluate a search filter,
the special nature of EXISTS filter was ignored.

It is now fixed.

Note: When a filter like "EXISTS(path1): path2=..." is used, both
path1 and path1/path2 has to be allowed in the authorization.
Please see role-search-user-assignment-targetRef.xml for an example.

This should resolve MID-7931.