Fix evaluation of modification policy constraints
The "items" property of modification policy constraints was
evaluated by simply looking at the list of modifications that
are to be applied. However, this did not work correctly when
(e.g.) the whole container (containing the item in question)
was added - see e.g. MID-7945. Similar situation could be when
the container is deleted.

This is now fixed by changing the approach altogether. Instead of
looking at the deltas, we directly compare "old" and "new" object
or assignment values regarding the respective item(s).

This should fix MID-7945.