Fix various problems in delayed shadow deletion
This commit re-implements a couple of earlier fixes made for 4.4 and 4.6
versions:
- dc42c96c015fff02d03db43bad9a3fcb659c200c
- faa835e627c36c6b165b798d13d0273c7198ec87
- 3780d62afe0d9011063ceebddbe3c93037528aaf

(The majority, but not all of them, though.)

The changes are:

1. Shadow lifecycle state is visible in LensProjectionContext.
(In a temporary way - an implementation different from 4.4.)
2. Reaped shadows are set to be "kept" (by ActivationProcessor).
3. Contexts with reaped shadows are not considered to be usable
when an account is to be created - i.e., a new projection context
is created in such situations.
4. Reaped accounts are treated specially also during context loading
(their linkRefs are not deleted, projections not marked as tombstone).
5. Projections for reaped accounts are not computed ("projected").
6. Reaped shadows are ignored during conflict checking.
7. PrimaryIdentifierValue index is cleared on entering "reaping"
state (and setting back again if the shadows goes back to live state).
8. ShadowManager.recordOperationException now updates the in-memory
version of the shadow.

Related to MID-8069.