Consider reaped account as being "almost gone"
Accounts in "reaping" lifecycle state were mostly treated like living
ones. However, this brings some problems when such an account is
re-created (see e.g. MID-8069): model tries to somehow revive such
account, leading to inconsistencies in its repository shadow.

This commit fixes the situation by introducing the following changes:

1. In clockwork, accounts in "reaping" state are considered just like
if they were gone. This means that when a new account for the object
type is to be created, a new projection context is allocated for it.
We no longer reuse the original context.

2. Accounts in this state have their primaryIdentifierValue cleared.
This allows us to create a "live" version of such account.

3. We ignore accounts being reaped during constraint violation checks
(just like we do for accounts that are gone).

4. Projection of accounts being reaped (in clockwork) is skipped.

This should resolve MID-8069.