Fix archetype handling (mainly for linked objects)
The code for determination of focus archetype in findLinkedSource,
findLinkedTarget and related methods was too simplistic. It failed
(in different ways) in situations when focus archetype was changed.

Changes in this commit:

1. Code for links resolution has been fixed. Although still not
perfect, it now looks for link definitions in new, current, and
old version of the focus; to cover definitions in both old and new
archetypes. This will work unless there are conflicting link
definitions in old and new archetypes.

2. The algorithm for archetype determination (in ArchetypeDeterminer)
has been simplified by ignoring values of archetypeRef. We kept
considering these values for a long time, mainly because originally
that was the primary mean of defining the archetype. But currently it
causes more troubles than benefits. So, from now on we look only at
archetype assignments in this class. (There are still some places
that look on archetypeRef, but they should NOT be used during clockwork,
when there may be inconsistencies between archetypeRef and assignments.)

Related to this, ObjectTypeUtil.hasArchetype was renamed to
ObjectTypeUtil.hasArchetypeRef, to emphasize the fact it uses
archetypeRef values to determine the result.

3. Determination of "source-linked" objects failed when the focus
OID was null. This is now fixed.

This should fix MID-8060 and MID-8059.