Check indirect assignments during authentication
Targets mentioned in "requireAssignmentTarget" were checked among
direct ones. This might be inconvenient for deployments that use role
inducements. Moreover, it did not take validity and conditions into
account. (The actual implementation also ignored relations.)

This is now fixed, and the matching is done on roleMembershipRef.
OID, type, and relation are checked. Resource assignments are not
supported.

!!! This commit can be considered a behavior change. !!!

This fixes MID-8123.

Unrelated change: Added PCV ID attribute to containers in
common-security-3.xsd. This should fix XSD compliance in MidPoint
Studio.