Clarify situation with GUI-started tasks
The tasks created by GUI on behalf of the user were not clearly defined:
the default End User role claimed an authorization is needed for them,
but in fact it was not. (Except for tasks for the custom actions
feature.)

This commit keeps the original behavior, and does the following changes:

1. The (unneeded) "self-owned-task-add-execute-changes" authorization
   was removed from 040-role-enduser.xml.
2. Introduced ModelInteractionService#switchToBackground method.
3. Two unused task-related methods were removed:
   - WebModelServiceUtils#runTask,
   - WebComponentUtil#createIterativeChangeExecutionTask.

Related to MID-8114.

Unrelated changes:

- Fixed ColumnUtils#findPropertyInResourceSet method.
- Removed unneeded RootTaskLoader.