Stop caching connectors with ad-hoc configuration
When a connector is instantiated for a "custom" resource object
(i.e., one that was not retrieved from repository, but provided to
ProvisioningService by the caller), we no longer put it into the
connector instance cache (MID-8020).

Also, this commit reorders initial test methods in AbstractLdapTest
in conntests to the state where they used to fail exactly because
the above bug. So, this effectively reverts point #1
in dd5702539a21b1a9cc57b4a69b135d25c03f2c48.

Resolves MID-8020.