Do not check autz when req roles for more users
We decided to stop requiring task-related authorizations for tasks that
are started autonomously by GUI.

The task created when requesting roles for more users was the last such
task (besides ones started by "custom actions") in midPoint 4.4. Its
submission was fixed to stop requiring authorizations in this commit.

This resolves MID-8114.