Ignore correlation state if correlation is over
In order to support asynchronous correlation, the correlation state was
introduced into shadows in 4.5. This state was used e.g. to continue in
the correlation process after the respective correlation case was
resolved.

However, as a side effect, the correlation on a given shadow was never
re-attempted: subsequent correlation processed provided always the
stored result.

This is now fixed:

1. On correlation re-attempt, the state is ignored (and cleared) if
the previous correlation was already ended.

2. On correlation case resolution, we no longer mark the whole
correlation as ended. It is marked as such only after the actual
correlation is run (following the case completion), taking the results
from the case into account.

This should resolve MID-8223.