Fix displaying projections
Two unrelated fixes are here:

1) An attempt to display unreadable projection failed with 500,
referring to inability to get a value from empty list. This should
be fixed by a change in PrismContainerWrapperImpl#findValue.

This could be causing MID-8136 (original stack trace) and maybe
MID-8218.

2) Triggers in shadows were copied onto themselves, creating collisions
in PCV IDs. The fix is dubious, because the whole code that copies
the triggers onto themselves is questionable. But, at least, the
exception is not thrown now.

This should resolve the issue in re-opened MID-8136. (Reproducible by
opening any projection that have some triggers on it.)