Improve ProvisioningService.getObject javadoc
In order to specify the behavior for various development and execution
modes, let us specify the "basic" behavior. This is done in this commit.

Also, the readability of the code was slightly improved.

No functional changes here, except for fixing the "raw" mode processing
when getting the shadows. We no longer compute shadow lifecycle state
nor futurize it in that mode. (See also MID-7419.)