Mark shadows as simulated also on creation
Shadows are now marked as "simulated" also when they are created
(using ProvisioningService#addObject method call).

Other changes:
- Added experimental support for querying resource attributes
in the QueryBuilder.
- Removed OperationResult from ResourceOperationDescription.
(It was quite messy, as it required the result to be completed before
the respective event is emitted; which is problematic. And it is not
necessary to have full OperationResult there.)
- Kind is no longer set in shadows being created. (It has no meaning
without the intent, after all.)
- Generally, the provisioning-impl code was slightly simplified.