Simplify ShadowManager and "shadows" pkg code
ShadowManager as a class no longer exists. It was a mere facade
lately, and all it did was forwarding method calls. Now it was replaced
by three classes: ShadowFinder, ShadowCreator, and ShadowUpdater which
collectively provide the shadow management functionality to "shadows"
package.

Some other parts in the "shadows" package were cleaned up as well.