Fix occasional failing of TestProgressReporting
The task used to fail on fast machines: it simply processed all 400
pre-created objects, so buckets after #40 were all empty. Hence,
the equation "# of buckets x size of bucket = # of objects" no longer
held.

This is now fixed by increasing the number of objects (so that
all buckets are now full), and decreasing the time available for task
run - so the task will not finish in the time available.