Introduce shadow death timestamp
Handling of dead shadows was not consistent: we marked them as dead
at defined places, and if this was on the right occasion (e.g. during
reconciliation run or with discovery enabled), we invoked appropriate
synchronization reaction. At other places, the shadow could be marked
dead without invoking any synchronization code. Moreover, some hacks
had to be done, namely avoiding marking shadows as dead when in
simulation mode. And still, the solution was not 100% correct:
e.g. it broke down when thresholds were in action (see
TestThresholds.test430ReconcileDeleteAllSimulateExecute).

So, in this commit, we introduced deathTimestamp property in shadow.
It is used (along with fullSynchronizationTimestamp) to determine if the
shadow has to be processed during 3rd part of the reconciliation.
We also no longer skip marking shadows as dead in simulation mode.

There are some induced changes:

- Full synchronization timestamp is now set only AFTER the
synchronization is executed; and only if the result is not FATAL_ERROR.
(This does not apply to synchronizationSituationDescription.)
- The DELETED sync reaction is run only if there is no "replacement"
live projection for particular user. This is to avoid problems of e.g.
account renames or delete+add operations (TestUuid).

Other changes:

- Task execution mode is set in LocalActivityRun according to the
activity execution mode.
- The "simulate" flag was removed from some provisioning events. It is
superseded by the task execution mode.
- Unneeded marking shadow as dead in ProjectionsLoadOperation is gone.
- New event-based task waiting mechanism in tests was fixed.

Work in progress. Some tests may fail.