Start implementing "simulated execution"
This is the first attempt at implementing "simulated execution"
at the model (clockwork) level.

Some ideas:

1. Not sure how to report simulated operations. One option is
to use the same mechanism as we have for audit (i.e. just like we
audit executed deltas, we could "audit" simulated ones). The drawback
is that some lower-level deltas could not be captured in this
way. So, as a quick hack, we use preliminary change listener bound
to the Task object.

2. Simulated execution of ADD, MODIFY, and DELETE operations is
implemented in DeltaExecution class. There are other places where such
redirection should be done, or at least researched, though.

3. Rotting of lens context had to be adapted, see the relevant changes.
To compensate for re-loading the changed objects from the repository,
ElementState#updateAfterExecution method had to be hacked a bit.

4. Of course, operation execution recording is disabled for simulations.
Auditing is disabled as well.

EXPERIMENTAL CODE. Brutal hacking for now.