Replace searchShadowOwner with equivalent search
The original searchShadowOwner execution was not cacheable
in global nor in local repository caches, resulting in excessive
number of search operations e.g. when authorizations were present.

Now the method is replaced by equivalent searchObjects call
at the level of RepositoryService, so that caching should be employed.
(Note that the translation to regular search operation was there
internally in SqaleRepositoryService implementation perhaps from the
day one.)