Providing stateBefore/stateAfter to sim manager
Here we had to rework the mechanism of passing info to simulation
result manager: instead of passing each delta as soon as it's
produced, we use information from LensElementContext at the end
of the clockwork. This way we have the full "object old" information
that would not be available otherwise.

Work in progress. The full info is not stored into the repo now;
and the deltas are not retrieved either.