Add dev mode for roles and assignments/inducements
This is an attempt to introduce "development mode" evaluation of roles,
assignments and inducements: in development mode the lifecycle state of
"proposed" should be among ones that are taken into account. (Later,
"deprecated" should be ignored in that mode.)

However, there is a collision with existing working of the lifecycle
state model. This should be resolved somehow.

Also, this commit prepares for mapping and object template lifecycle
state definition.