Backport resource delta applyDefinition fix
This is an attempt to backport the fix for MID-7918
(992d2739cf2ae48ee8ead190ef93279c68b8f58a). The problem is that
the fix itself refers to the code that was heavily refactored
during the development of 4.5-4.6.

Hence, here we practically replace parts of the legacy code with their
modern equivalents: ConnectorSpec class and the code for applyDefinition
method related to resource definition deltas.