Implement event tags PoC more conservatively
This is an experiment to implement so-called event tags without the need
to add new ObjectType (TagType). Here the tags are represented by plain
string values (e.g. URIs). The rest is the same as in
d9c1b5d4446bfd5a069295f0ffb533503e54ed4e:

1. Each policy rule has a set of tags it induces. These are (by default)
not written anywhere, just collected for the purposes of simulation and
auditing.
2. In the former case, these event tags are stored as metric identifiers
in SimulationResultProcessedObjectType. (Will be probably renamed.)

Other changes:

- The AggregatedObjectProcessingListener (temporary name) was adapted
to work with event tags.
- The XSD for SimulationResultProcessedObjectType was fixed, and the
creation of these objects was improved (see ProcessedObject class).

Work in progress. The final form of event tags should be decided.
It may be a string (OID, URI) or QName.