Add basic event tags
Here we attempt to create the standard set of event tags that will be
used to mark operations in simulation mode (and later audit events in
real execution mode).

In particular:

1. Tag archetypes were created: event tag, policy situation.
2. First 9 event tags were added.
3. New "custom" policy constraint was introduced.
4. The basic "Standard functions" library was created; even if it is
not currently used for tag policy rules because of performance and
usability questions.
5. Expression factory is now set at various places, fixing missing
support for "function" evaluator (e.g. in policy constraints).