Remove TagType support from H2 repository
The support is not necessary in 4.7, because we plan to use tags solely
for simulations now (and those will not be supported by generic repo,
either). So it is more efficient to drop preliminary H2 support than
to introduce support for the other databases.

Unrelated change:
 - AbstractTestResource#get now returns a copy of the stored object,
to avoid unintentional modification by the client code.