Improve simulations-related schema
Major schema changes:

1. SimulationDefinitionType was introduced. The system config
now contains a list of these definitions (to be selected from for the
tasks). A task can contain a custom simulation definition as well.

2. SimulationResultType has now start and end timestamps
and its definition.

Other changes:

- Simulation result is closed when the activity finishes: the end
timestamp is recorded and the metric values are computed. (Not working
quite well with distributed and delegated activities for now.)

- Minor code improvements: documenting and renaming things.