Native repo: LEFT JOIN is used when ordering by ref target item
This allows ordering by the target item for reference search where
it makes sense even for multi-value refs, because each ref value has
its own result row.
Target conditions are still applied by EXISTS (semi-join), but that's
not a correctness problem; possible improvements mentioned in comments.
Also, unique aliases are now not checked against JOINs of the Querydsl
query, but tracked explicitly in a Set in the root query context.