Compute simulation metrics incrementally
This commit switches computation of simulation metrics from "at the end
of simulation" (time-consuming, late, failing on suspend/resume) to
"at the end of each bucket".

To support this, a new concept of "simulation result transaction" was
introduced.

Other improvements:

1. LensElementContext is now available for custom metrics computation.
2. Metric values summation and (partial/total) dimension collapsing
are now implemented correctly.

Work in progress. E.g. simulation result transactions still have to be
implemented in the repository.