Fix simulations for perpetual state activities
Activities that have perpetual state (e.g. LiveSync) will now create
a new simulation result on each individual realization (~ run). This is
to avoid writing to already closed simulation results. It may be
reconsidered later; the official statement is that simulations are
supported for LS tasks, but only for a single execution.