Block provisioning changes for simulation tasks
The provisioning will refuse to carry out any "state-changing"
operations, including script execution. This is done at various levels,
to avoid e.g. recording such operations as pending, with the danger
of later execution.

These are blocked at:

- shadows facade level (i.e. almost at provisioning API level),
- ResourceObjectConverter (i.e. when UCF API is called),
- implementations of ConnectorInstance (i.e. right within ucf-xxx-impl).

At all of these places, an exception is thrown, i.e. the caller is
responsible to avoid making such calls. A special case is the code that
deals with the shadow refresh, as it can be invoked quite autonomously.
(This may be reconsidered.)