Support "report-only mode" for attributes (PoC)
This commit introduces "report" and "ignore" modes for resource object
attributes. The former causes a delta to be written to a simulation
result each time this particular attribute would be changed by the
operation. The latter causes all changes of the attribute to be ignored.

Proof of concept quality; meant as a basis for discussion.