Support LC state in attributes and associations
Now it's possible to define the lifecycle state at the level of
a [resource object] attribute and association. This turns off all
inbound and outbound mappings' evaluation, as well as the reconciliation
of that attribute/association with the value on resource. (Just like
setting "IGNORE" processing at the level of model.)

Also, the LC state of "deprecated" is now generally excluded from
the processing in the development mode.

Unrelated change: The ignore/non-readable flags are now checked also
for pre-clockwork inbound processing.

Work in progress.