Fix handling reportDataRef in distributed reports
Distributed reports created two ReportDataOutput objects, by mistake.
One at the beginning of the execution, to serve as a parent for partial
data objects, and another one at the end, to hold the aggregated data.

This is now fixed. At the end, we update the pre-existing one with the
final information.

This may be a slightly incompatible change, as it requires "MODIFY"
authorizations in addition to "ADD" ones in order to run distributed
report exports.

Plus some cosmetic/readability improvements.