Refactor policy rules processing
The immediate goal of this commit is to enable evaluation of policy
rules even if there is no focus present, e.g. when unlinked shadow is
deleted.

(See TestModelServiceContract#test126DeleteUnlinkedAccountSimulated.)

The package com.evolveum.midpoint.model.impl.lens.projector.policy and
its interface to rest of Lens had to be cleaned up and re-organized.
The PolicyRuleProcessor is now the main interface to the whole package.

Work in progress. Some tests may fail.