Fix correlation processing
The last start/end timestamp of correlation processing was not updated;
this was broken recently in 9a254359bfea939fd0219bb344cc09c310ccf855,
when the clearing was removed to facilitate the simulations.

Here we introduce at least setting of start and end timestamps,
hoping it will be enough.

This should fix TestFirstSteps.