Fix shadow management simulation
1. Fixed deltas issued by correlation processing: introducing estimated
old values, optimizing deltas if no real changes are there.

2. Fixed deltas issued by synchronization service:
a) optimizing deltas if no real changes are there,
b) resetting sync situation to null for protected objects,
c) overwriting "unknown" classification even if synchronization is
excluded (we do that in provisioning module anyway, so this is a code
simplification),
d) updating synchronization timestamp even in shadow mgmt simulation
mode (this is needed for 3rd reconciliation stage).

3. Removed check for protected shadows in Synchronizer, because the same
functionality is present in SynchronizationServiceImpl.

4. Adapted and extended TestFirstSteps.