Fix assigned projections with focus iteration
When the assigned resource object constructions are evaluated repeatedly
(for example, when focus iteration is engaged), midPoint may try to
fetch not-yet-existing projection. It is because of limitations of the
current LensProjectionContext#isAdd implementation (see the javadoc
added).

Either we would improve that method, or - as a quicker albeit maybe not
so comprehensive fix - we simply avoid loading projections without OID
when evaluating such constructions. The latter approach was selected
for this commit.

This resolves MID-8569.