Fix one-sided exclusion policy rules
Support for exclusion policy rules that are defined on one side only
(as specified in the documentation) was only partial: enforcement was
working but approvals and state recording ware not.

This is now fixed.

It should resolve MID-8269 and (positively) influence related GUI
issues: MID-8243, MID-8251, and probably others.

Other changes:
- Fixed recording of the state related to pruned assignments (MID-7123).