Improve Linked Objects scenario #5
1. Exception logging for "assignment targetRef object not found" was
weakened: instead of error(..) (with full stack trace) we use
logException(..) that provides the full stack trace only for DEBUG level
and if the assignment is being deleted, the message is hidden completely
(visible only on DEBUG level).

This may be a little questionable, but should cause no harm overall.

2. A working sample for asynchronous scripting rule execution was added.

3. When asynchronous policy rule execution is requested, the task OID
storage in the result is fixed (the "taskOid:" prefix was not used).

Related to MID-8366.