Optimize approverRelation "OR" query
The "OR" REF query over multiple relations (specified by
`approverRelation` property) was inefficient on some databases.
(Especially for old generic repo, where multiple left joins were
used in such cases.)

It is now replaced by a set of independent queries, with the results
merged afterwards.

We assume that there will not be a large number of relations used, so
the number of queries will be low.

This resolves MID-8134.