Improve "one-sided exclusion" policy rules
When a "one-side" exclusion policy rule triggers on the assignment on
which it is defined, this information is propagated onto the conflicting
assignment - this was implemented in
a296429170e721e61d525e2cbc8acd1a654f7e7c.

However, the problem is when there are multiple conflicting assignments.
Since the creation of policy rules, only one trigger was produced for
each constraint. Hence, only a single conflicting assigment could be
"marked" by the rule after the evaluation.

This is now fixed. All relevant triggers are processed. However, not all
of them are relevant to all conflicting assignments! So, a new mechanism
had to be conceived to handle this - see AssociatedPolicyRule interface
and ForeignPolicyRuleImpl class.

This resolves MID-8251.