Disable "account/default" hack for synchronization
In order to support various legacy and "barely correct" configurations,
midPoint - at several places - considers ri:AccountObjectClass shadows
as ones with the "account/default" type even if formally such type is
undefined or these account do not match the criteria.

Such approach is generally questionable, and should be reconsidered in
the future.

This commit removes it from the place, where it obviously should not be
applied: when synchronizing a resource object.

This resolves MID-8516.