Fix wrong type of returned shadow
When retrieving shadows in "simulated shadow management" mode, the
new classification was stored into the simulation result, but not
returned in retrieved shadows. This prevented the correct kind/intent
from being displayed in resource content preview in GUI.

This commit fixes that.

It resolves the first part of MID-8613.

(Also fixes the report archetypes handling in two simulation tests
after recent changes.)