Stop applying authorizations onto definitions
In order to improve the performance of model.getObject and similar
method calls, we decided to reduce the processing done on objects being
returned.

The first candidate is the application of authorizations on object
definitions. It looks like these need not be done. (The
getEditObjectDefinition method should be sufficient for this.)
So, this commit eliminates this processing.

Work in progress. To be tested.

There are other (mainly readability) changes. The following experimental
auxiliary classes were created:

- ConfigErrorReporter
- ParsedGetOperationOptions