Turn off schema refinements application by default
This is an experimental change: In order to improve the performance
of get/search model API calls, we now turned off the application of
schema refinements (from object templates) altogether.

The tests that checked for this functionality were adapted. The question
is whether there are some implicit assumptions that these refined
definitions are applied.